Adjusted Permission-Requesting corresponding to Johannes' Feedback: - onRequestPermissionsResult implemented only in RequestPermissionsActivity - Moved requesting for permissions on Restart of the App into the onRestart-Method of the RequestPermissionsActivity
